Bug 702810: fix fapi buffer size parameter types
The buffer size parameters for the fapi callbacks used to be unsigned short,
but were changed to ints a while back. The calls for the (g)subr data in
Type 2/CFF fonts hadn't had the casts to ushort removed, causing the values
to be trucated.

There was also a Postscript/FAPI internal function that had not had the
buffer length parameter changed to int.

+++ psi/zfapi.c
@@ -1704,7 +1704,7 @@ decode_bytes(byte *p, const byte *s, int l, int lenIV)
 
 static int
 get_type1_data(gs_fapi_font *ff, const ref *type1string,
-               byte *buf, ushort buf_length)
+               byte *buf, int buf_length)
 {
     gs_font_type1 *pfont = (gs_font_type1 *) ff->client_font_data;
     int lenIV = max(pfont->data.lenIV, 0);
